% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\docType{data}
\name{bank}
\alias{bank}
\title{Banknote Authentication}
\format{
A data frame with 1372 rows and 5 variables.
\itemize{
 \item target: target variable, class.
 \item variance: variance of Wavelet Transformed image.
 \item skewness: skewness of Wavelet Transformed image.
 \item curtosis: curtosis of Wavelet Transformed image.
 \item entropy: entropy of image.
}
}
\source{
<https://doi.org/10.24432/C55P57#>
}
\usage{
bank
}
\description{
Data were extracted from images that were taken for the evaluation of an
authentication procedure for banknotes.
}
\details{
\itemize{
 \item author: Lohweg, Volker.
 \item year: 2012.
 \item publisher: UCI Machine Learning Repository.
 \item doi: 10.24432/C55P57.
 \item url: https://doi.org/10.24432/C55P57#.
}
}
\keyword{datasets}
