% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brcancer_diag.R
\docType{data}
\name{brcancer_diag}
\alias{brcancer_diag}
\title{Breast Cancer Wisconsin (Diagnostic)}
\format{
A data frame with 569 rows and 31 variables.
#' \itemize{
 \item target: Diagnosis.
 \item 30 features, for more details see url.
 }
}
\source{
<https://doi.org/10.24432/C5DW2B>
}
\usage{
brcancer_diag
}
\description{
Features are computed from a digitized image of a fine needle aspirate (FNA) of a breast mass.
They describe characteristics of the cell nuclei present in the image.
A few of the images can be found at http://www.cs.wisc.edu/~street/images/.
}
\details{
\itemize{
 \item author: Wolberg, William; Mangasarian, Olvi; Street, Nick; Street, W.
 \item year: 1993.
 \item publisher: UCI Machine Learning Repository.
 \item doi: 10.24432/C5DW2B.
 \item url: https://doi.org/10.24432/C5DW2B.
 }
}
\keyword{datasets}
