% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brcancer_prog.R
\docType{data}
\name{brcancer_prog}
\alias{brcancer_prog}
\title{Breast Cancer Wisconsin (Prognostic)}
\format{
A data frame with 198 rows and 32 variables (for more details see url).
\itemize{
 \item target: Outcome, recur or non-recur.
 \item 31 features, for more details see url.
 }
}
\source{
<https://doi.org/10.24432/C5GK50>
}
\usage{
brcancer_prog
}
\description{
Each record represents follow-up data for one breast cancer case.
These are consecutive patients seen by Dr. Wolberg since 1984,
and include only those cases exhibiting invasive breast cancer
and no evidence of distant metastases at the time of diagnosis.
}
\details{
\itemize{
\item author: Wolberg, William; Street, W.; Mangasarian, Olvi.
\item year: 1995.
\item publisher: UCI Machine Learning Repository.
\item doi: 10.24432/C5GK50.
\item url: https://doi.org/10.24432/C5GK50.
}
}
\keyword{datasets}
