% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{gmm_classifier}
\alias{gmm_classifier}
\title{Classifier using GMM chains}
\usage{
gmm_classifier(
  x,
  y,
  n,
  thres,
  proc_name = "train",
  graph_dir = sprintf("\%s/graphics", getwd()),
  model_gmm = NULL,
  gmm_args = list(num_chains = 3L, chain_length = 3L, draw_de = TRUE),
  ...
)
}
\arguments{
\item{x}{Predictors in data frame.}

\item{y}{Target variable.}

\item{n}{size of data, length(y).}

\item{thres}{Accuracy threshold.}

\item{proc_name}{In training or testing phase.}

\item{graph_dir}{Directory for storing graphics.}

\item{model_gmm}{Trained GMM classifier.}

\item{gmm_args}{A list of parameters for GMM chains.
\itemize{
 \item num_chains: Number of GMM chains.
 \item chain_length: Maximum iterations.
 \item draw_de: If to draw GMM based density estimator of the classifier in GMM chain.
}}
}
\value{
A list with elements.
\itemize{
 \item row_class_gmm: classification info by GMM Classifier of all rows in data frame.
 \item model_gmm: GMM Classifier trained.
}
}
\description{
Classifier using GMM chains
}
\keyword{internal}
