% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leukemia.R
\docType{data}
\name{Leukemia}
\alias{Leukemia}
\title{Prediction of acute leukaemia subtypes}
\format{
A data frame with 121 rows and 15 variables.
}
\source{
<https://doi.org/10.1016/S2589-7500(24)00044-X>
}
\usage{
Leukemia
}
\description{
Data from journal paper:
Evaluation of a machine-learning model based on laboratory parameters for the prediction of
acute leukaemia subtypes: a multicentre model development and validation study in France.
121 patients diagnosed with AML, APL, or ALL were included.
}
\details{
\itemize{
 \item journal: The Lancet Digital Health.
 \item volume: 6.
 \item number: 5.
 \item pages: e323-e333.
 \item year: 2024.
 \item doi: https://doi.org/10.1016/S2589-7500(24)00044-X.
 \item author: Vincent Alcazer, Grégoire Le Meur, Marie Roccon, et.al.
 }
}
\keyword{datasets}
