% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_run}
\alias{model_run}
\title{Model run}
\usage{
model_run(
  dat,
  dat_nm = "data",
  target_var = "target",
  features_ignored_gmm = NULL,
  test_size = 0.3,
  stratify = TRUE,
  model_names = c("lr", "gmm"),
  thres = 0.98,
  min_thres = 0.92,
  rep_labels = 1L:10L,
  seed = 2025,
  model_saved = FALSE,
  result_dir = sprintf("\%s/result/\%s", getwd(), dat_nm),
  graph_dir = sprintf("\%s/graphics/\%s", getwd(), dat_nm),
  opt_metric = "Accuracy",
  opt_phase = "test",
  lr_params = list(penalty = FALSE, strict_selection = FALSE, test_type = "Wald",
    sig_level = 0.05, min_features = 5L),
  gmm_args = list(num_chains = 3L, chain_length = 3L),
  ...
)
}
\arguments{
\item{dat}{Data to be classified in data frame.}

\item{dat_nm}{Name of data, default "data".}

\item{target_var}{Name of target variable, default "target".}

\item{features_ignored_gmm}{Features ignored by GMM Classifier, default NULL.}

\item{test_size}{Proportion of test set, default 0.3.}

\item{model_names}{Model names, default linear regression (lr) and GMM classifier (gmm).}

\item{thres}{Accuracy threshold, default 0.98.}

\item{min_thres}{accuracy rate needed, default 0.92.}

\item{rep_labels}{Labels for the repeated experiments.}

\item{seed}{Seed for generating random number.}

\item{model_saved}{If trained model to be saved for stability test, default FALSE.}

\item{result_dir}{Directory where results are stored.}

\item{graph_dir}{Directory where graphics are stored.}

\item{opt_metric}{Metric based on which the optimal chain is selected, default "Accuracy".}

\item{opt_phase}{Which phase is used for selecting optimal chain, default "test".}

\item{lr_params}{A list of parameters for Logistic Regression.
\itemize{
 \item penalty: If select features using LASSO, default FALSE.
 \item strict_selection: If select features strictly, e.g. features with non-zero
coefficients in LASSO model or features with p-value <= sig_level in logistic regression,
default FALSE.
 \item test_type: Significant test for selecting features from LR model, default "Wald".
When set to "Wald", a normal approximation is used,
otherwise a more accurate likelihood ratio test is carried out.
 \item sig_level: Significant level, default 0.05.
 \item min_features: Minimal number of features to be selected, default 5.
}}

\item{gmm_args}{A list of parameters for GMM Chains.
\itemize{
 \item num_chains: Number of GMM chains.
 \item chain_length: Maximum iterations.
}}

\item{strafity}{If data is split in a stratified fashion, default TRUE.}
}
\value{
A list of elements
\itemize{
 \item metric_rec: Model metrics like accuracy, precision, recall, f1, roc in data frame.
 \item template_rec: Feature templates models returned.
 \item result_dir: Directory where results are stored.
}
}
\description{
Run LR-GMM model length(rep_labels), default 10 times on given data with random train/test split.
}
\examples{

################### demo 1: Diagnosis of breast cancer
## Binary classification
library(gmmcc)
data(brcancer_diag)
model_result <- model_run(brcancer_diag, dat_nm = "brcancer_diag",
                         lr_params = list(penalty = TRUE, min_features = 5),
                         model_saved = TRUE, opt_metric = "AUC")
# accuracy rates from training phase vs. testing phase
accuracy <- model_result$metric_rec$Accuracy
phase <- model_result$metric_rec$phase
boxplot(accuracy ~ phase)

################### demo 2: Sales Forecasting
## Multiclass classification
data(sale)
model_result <- model_run(sale, dat_nm = "sale",
                          gmm_args = list(num_chains = 3),
                          model_saved = TRUE, opt_metric = "AUC")
# accuracy rates from training phase vs. testing phase
accuracy <- model_result$metric_rec$Accuracy
phase <- model_result$metric_rec$phase
boxplot(accuracy ~ phase)

}
