% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_stab_test}
\alias{model_stab_test}
\title{Model stability test on different train/test sets}
\usage{
model_stab_test(
  dat,
  dat_nm = "data",
  target_var = "target",
  rep = 10,
  seed = 2025,
  test_sizes = c(0.2, 0.3, 0.4),
  stratify = TRUE,
  features_ignored_gmm = NULL,
  result_dir = paste0(getwd(), "/result"),
  model_chosen,
  no_chain,
  ...
)
}
\arguments{
\item{dat}{Data to be classified in data frame.}

\item{dat_nm}{Name of data, default "data".}

\item{target_var}{Name of target variable, default "target".}

\item{rep}{Number of repeat experiments, default 10.}

\item{seed}{Seed for generating random number.}

\item{test_sizes}{Different test sizes, default c(0.2, 0.3, 0.4).}

\item{features_ignored_gmm}{Features ignored by GMM Classifier.}

\item{result_dir}{Directory where results are stored.}

\item{model_chosen}{Model to be tested.}

\item{no_chain}{Which chain to be tested.}

\item{strafity}{If data is split in a stratified fashion, default TRUE.}
}
\value{
stab_test_result Result of stability test.
}
\description{
Model stability test on different train/test sets
}
\keyword{internal}
