% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_test}
\alias{model_test}
\title{Model testing}
\usage{
model_test(
  x,
  y,
  features_ignored_gmm = NULL,
  thres = 0.98,
  model_trained = NULL,
  graph_dir = sprintf("\%s/graphics", getwd()),
  ...
)
}
\arguments{
\item{x}{Predictors in data frame.}

\item{y}{Target variable.}

\item{features_ignored_gmm}{Features ignored by GMM Classifier, default NULL.}

\item{thres}{Accuracy threshold, default 0.98.}

\item{model_trained}{Trained model from training phase.}

\item{graph_dir}{Directory where graphics are stored.}
}
\value{
row_class Classification info of all rows in data frame
}
\description{
Model testing
}
\keyword{internal}
