% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_train}
\alias{model_train}
\title{Model training}
\usage{
model_train(
  x,
  y,
  features_ignored_gmm = NULL,
  model_names = c("lr", "gmm"),
  thres = 0.98,
  min_thres = 0.92,
  graph_dir = sprintf("\%s/graphics", getwd()),
  lr_params = list(penalty = FALSE, strict_selection = FALSE, test_type = "Wald",
    sig_level = 0.05, min_features = 5L),
  gmm_args = list(num_chains = 3L, chain_length = 3L),
  ...
)
}
\arguments{
\item{x}{Predictors in data frame.}

\item{y}{Target variable.}

\item{features_ignored_gmm}{Features ignored by GMM Classifier, default NULL.}

\item{model_names}{Model names, default linear regression (lr) and GMM classifier (gmm).}

\item{thres}{Accuracy threshold, default 0.98.}

\item{min_thres}{accuracy rate needed, default 0.92.}

\item{graph_dir}{Directory where graphics are stored.}

\item{lr_params}{A list of parameters for Logistic Regression model.
\itemize{
 \item penalty: If select features using LASSO, default FALSE.
 \item strict_selection: If select features strictly, e.g. features with non-zero
 coefficients in LASSO model or features with p-value <= sig_level in logistic regression,
default FALSE.
 \item test_type: Significant test for selecting features from LR model, default "Wald".
When set to "Wald", a normal approximation is used,
otherwise a more accurate likelihood ratio test is carried out.
 \item sig_level: Significant level, default 0.05.
 \item min_features: Minimal number of features to be selected, default 5L.
}}

\item{gmm_args}{A list of parameters for GMM chains.
\itemize{
 \item num_chains: Number of GMM chains.
 \item chain_length: Maximum iterations.
}}

\item{...}{dots argument.}
}
\value{
A list of elements.
\itemize{
 \item model_trained: Mixture model trained.
 \item row_class: Classification info of all rows in data frame.
}
}
\description{
Model training
}
\keyword{internal}
