% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sale.R
\docType{data}
\name{sale}
\alias{sale}
\title{Wholesale Customers}
\format{
A data frame with 440 rows and 7 variables.
\itemize{
 \item target: Region, target variable.
 \item Channel: Feature, Categorical type.
 \item Fresh: Feature, Integer type.
 \item Milk: Feature, Integer type.
 \item Grocery: Feature, Integer type.
 \item Frozen: Feature, Integer type.
 \item Detergents_Paper: Feature, Integer type.
 \item Delicassen: Feature, Integer type.
 }
}
\source{
<https://doi.org/10.24432/C5030X>
}
\usage{
sale
}
\description{
The data set refers to clients of a wholesale distributor.
It includes the annual spending in monetary units (m.u.) on diverse product categories.
}
\details{
\itemize{
 \item author: Cardoso, Margarida.
 \item year: 2013.
 \item publisher: UCI Machine Learning Repository.
 \item doi: 10.24432/C5030X.
 \item url: https://doi.org/10.24432/C5030X.
 }
}
\keyword{datasets}
