% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{stab_test_run}
\alias{stab_test_run}
\title{Run model stability test}
\usage{
stab_test_run(
  dat,
  dat_nm,
  model_result,
  ref_phase = "test",
  eval_inds = list(eval1 = "Accuracy", eval2 = "AUC"),
  model_ref = NULL,
  ...
)
}
\arguments{
\item{dat}{Data to be classified in data frame.}

\item{dat_nm}{Name of data, default "data".}

\item{model_result}{Result from model_run.}

\item{ref_phase}{Reference phase, only "test" supported.}

\item{eval_inds}{Evaluation indicators used to choose reference model.}

\item{model_ref}{Reference model, default NULL.}
}
\value{
A list of elements
\itemize{
 \item result: Result of stability tests.
 \item template_chosen: Feature template of the chosen model.
}
}
\description{
Run stability test on the best model from the result of model_run.
}
\examples{

################### demo 1: Diagnosis of breast cancer
## Binary classification
library(gmmcc)
data(brcancer_diag)
model_result <- model_run(brcancer_diag, dat_nm = "brcancer_diag",
                         lr_params = list(penalty = TRUE, min_features = 5),
                         model_saved = TRUE, opt_metric = "AUC")
# model stability test
test_result <- stab_test_run(brcancer_diag, "brcancer_diag", model_result)
accuracy <- test_result$result$Accuracy
test_size <- test_result$result$test_size
boxplot(accuracy ~ test_size)

################### demo 2: Sales Forecasting
## Multiclass classification
data(sale)
model_result <- model_run(sale, dat_nm = "sale",
                          gmm_args = list(num_chains = 3),
                          model_saved = TRUE, opt_metric = "AUC")
# model stability test
test_result <- stab_test_run(sale, "sale", model_result)
accuracy <- test_result$result$Accuracy
test_size <- test_result$result$test_size
boxplot(accuracy ~ test_size)

}
