% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{train_test_split}
\alias{train_test_split}
\title{To split data into train and test sets.}
\usage{
train_test_split(
  dat,
  test_size = 0.2,
  shuffle = TRUE,
  stratify = TRUE,
  target_var = "target",
  split_var = NULL,
  seed = NULL
)
}
\arguments{
\item{dat}{Data in data frame.}

\item{test_size}{Proportion of test set, default 0.3.}

\item{shuffle}{Whether or not to shuffle data before splitting, default TRUE.}

\item{stratify}{If data is split in a stratified fashion, default TRUE.}

\item{target_var}{Target variable.}

\item{split_var}{Stratified variable, default target_var.}

\item{seed}{Seed for shuffling.}
}
\value{
x_train, y_train, x_test, y_test, dat_train, dat_test, seeds
}
\description{
To split data into train and test sets.
}
\keyword{internal}
